+++ src/Components/API/index.ts
@@ -12,10 +12,8 @@ import { TApiContext } from "./ApiContext";
 import { Store } from "../KeyValueStore";
 import { getMultiActionItems } from "../Actions/MultiAction/util";
 import { isHouseInformationComponent } from "./houseInformation";
-import { isGraphqlApi } from "./graphql";
+import { isGraphqlApi, graphQLApiHandler } from "./graphql";
 import uuid from "uuid";
-import get from "lodash.get";
-import { GraphQLError, ExecutionResult } from "graphql";
 
 export const callApi = async (
   component: ApiComponent,
@@ -25,77 +23,8 @@ export const callApi = async (
   changePassage: (name: string) => void
 ) => {
   if (isGraphqlApi(component)) {
-    try {
-      const variables = component.data.variables.reduce((curr, variable) => {
-        switch (variable.as) {
-          case "string":
-            return { ...curr, [variable.key]: String(store[variable.from]) };
-          case "int":
-            return { ...curr, [variable.key]: parseInt(store[variable.from]) };
-        }
-
-        return { ...curr, [variable.key]: store[variable.from] };
-      }, {});
-
-      const handleErrors = (graphqlResult: ExecutionResult<any>) => {
-        const error = component.data.errors.find(error =>
-          graphqlResult
-            .errors!.map((graphQLError: GraphQLError) => {
-              if (error.contains) {
-                return graphQLError.message.includes(error.contains);
-              }
-
-              return true;
-            })
-            .includes(true)
-        );
-
-        if (error) {
-          changePassage(error.next.name);
-        }
-      };
-
-      const handleData = (graphqlResult: ExecutionResult<any>) => {
-        component.data.results.forEach(result => {
-          setValue(result.as, get(graphqlResult.data!, result.key));
-        });
-
-        changePassage(component.data.next.name);
-      };
-
-      if (component.data.query) {
-        const graphqlQueryResult = await apiContext.graphqlQuery(
-          component.data.query,
-          variables
-        );
-
-        if (graphqlQueryResult.data) {
-          handleData(graphqlQueryResult);
-        } else if (graphqlQueryResult.errors) {
-          handleErrors(graphqlQueryResult);
-        }
-
-        return;
-      } else if (component.data.mutation) {
-        const graphqlMutationResult = await apiContext.graphqlMutation(
-          component.data.mutation,
-          variables
-        );
-
-        if (graphqlMutationResult.data) {
-          handleData(graphqlMutationResult);
-        } else if (graphqlMutationResult.errors) {
-          handleErrors(graphqlMutationResult);
-        }
-
-        return;
-      }
-
-      throw new Error("neither mutation or query was passed to graphql api");
-    } catch (e) {
-      console.log(e);
-      return;
-    }
+    graphQLApiHandler(component, apiContext, store, setValue, changePassage);
+    return;
   }
 
   if (isPersonalInformationApiComponent(component)) {
